Adding new Twig visitors and workers
