+++ src/Visitor/Twig/Twig1Visitor.php
@@ -13,13 +13,12 @@ namespace Translation\Extractor\Visitor\Twig;
 
 use Translation\Extractor\Visitor\BaseVisitor;
 use Twig_Environment;
-use Twig_Node;
+use Twig_NodeInterface;
 
 /**
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
- * @deprecated Use Twig2Visitor
  */
-final class Twig2TranslationFilter extends BaseVisitor implements \Twig_NodeVisitorInterface
+final class Twig1Visitor extends BaseVisitor implements \Twig_NodeVisitorInterface
 {
     /**
      * @var WorkerTranslationFilter
@@ -28,17 +27,17 @@ final class Twig2TranslationFilter extends BaseVisitor implements \Twig_NodeVisi
 
     public function __construct()
     {
-        $this->worker = new WorkerTranslationBlock();
+        $this->worker = new Worker();
     }
 
-    public function enterNode(Twig_Node $node, Twig_Environment $env)
+    public function enterNode(Twig_NodeInterface $node, Twig_Environment $env)
     {
         return $this->worker->work($node, $this->collection, function () {
             return $this->getAbsoluteFilePath();
         });
     }
 
-    public function leaveNode(Twig_Node $node, Twig_Environment $env)
+    public function leaveNode(Twig_NodeInterface $node, Twig_Environment $env)
     {
         return $node;
     }
