+++ src/Visitor/Twig/Twig2TranslationBlock.php
@@ -17,6 +17,7 @@ use Twig_Node;
 
 /**
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
+ * @deprecated Use Twig2Visitor
  */
 final class Twig2TranslationBlock extends BaseVisitor implements \Twig_NodeVisitorInterface
 {
