--- tests/Resources/Github/issue_34.include.html.twig
@@ -1,5 +1,0 @@
-{% embed 'issue_34.include.html.twig'  %}
-    {% block content %}
-        {{ 'trans.issue_34'|trans }}</p>
-    {% endblock %}
-{% endembed %}
